Allow load of non-existing pattern files.
To allow pure usage of the runtime API/LUA to add/remove entries in Maps/ACLs we should be able to load non-existing patterns.
This make deploying easier as otherwise empty files needs to be created.

A notice is issued when this config option is used and every time a non-existing file is loaded.

Fixes: 2202
Signed-off-by: Daan van Gorkum <djvg@djvg.net>